+++ etna/packages/etna-js/upload/api/upload_api.jsx
@@ -1,6 +1,6 @@
-// Use the etna-js/upload/utils/fetch version
+// Use the etna-js/utils/base_fetch version
 //   instead of etna-js/utils/fetch.
-import { form_post, json_post } from '../utils/upload_fetch';
+import { form_post, json_post } from '../../utils/base_fetch';
 
 export const postAuthorizeUpload = (base_url, project_name, bucket_name, file_path) => {
   let request = { project_name, file_path, bucket_name };
