+++ etna/packages/etna-js/utils/fetch.js
@@ -1,14 +1,8 @@
 import downloadjs from 'downloadjs';
 import * as Cookies from './cookies';
+import {base_json_fetch, checkStatus, isJSON, form_post} from './base_fetch';
 
-export const checkStatus = (response) => {
-  let content = isJSON(response) ? response.json() : response.text();
-  if (response.status >= 200 && response.status < 300) {
-    return content;
-  } else {
-    throw content;
-  }
-};
+export {checkStatus, isJSON, form_post};
 
 export const parseJSON = (response) => {
   try {
@@ -28,9 +22,6 @@ export const generateDownload = (filename) => {
     return downloadjs(blob, filename, blob.type);
   };
 };
-
-export const isJSON = (response) =>
-  response.headers.get('Content-Type') == 'application/json';
 
 export const headers = (...types) => {
   let _headers = {};
@@ -70,29 +61,11 @@ export const postOpts = (body) => ({
 });
 export const deleteOpts = {method: 'DELETE', ...opts};
 
-export const json_fetch = (method) => (path, params) =>
-  fetch((CONFIG.baseURL || '') + path, {
-    method,
-    credentials: 'include',
-    headers: headers('json'),
-    ...(params && {body: JSON.stringify(params)})
-  }).then(checkStatus);
+export const json_fetch = base_json_fetch;
 
 export const json_get = json_fetch('GET');
 export const json_delete = json_fetch('DELETE');
 export const json_post = json_fetch('POST');
-
-export const form_post = (path, params, performCheckStatus = true) => {
-  let form = new FormData();
-
-  Object.keys(params).forEach((key) => form.append(key, params[key]));
-
-  return fetch(path, {
-    method: 'POST',
-    credentials: 'same-origin',
-    body: form
-  }).then(performCheckStatus ? checkStatus : (v) => v);
-};
 
 export const handleFetchError = (e) => {
   console.log(e);
